# Feature or Bug Fix
[Issue 3260602] SP connector batch authZ should handle site collection URLs correctly for authZ grouping of URLs to avoid intermittent indeterminate results.

The authorization failed for default.aspx page URL for a site, when the site collection for the web application was created using the "/sites/" path.

http://SharePointSiteURL:portno/sites/subsite/socialsite/default.aspx

# High level list of changes

o Modified code for 'GetServerRelativeWebUrl' function and added code to retrieve correct server relative URL for a site.
o The logic works for both the cases, i.e. the site collection created using "/" path or "/sites/" path.

# Specific changes


GSBulkAuthorization.asmx
-----------------------
o Modified code in GetServerRelativeWebUrl function, for extracting proper substring which will correspond to the server relative URL for a site's default.aspx page. Logic works even when the site collection is created using the "/sites/" path.