Move long running SQL statements from SQL to Java
This patch breaks up all potentially long running operations in SQL into
individual row-level operations, so that authn resolution to be scheduled during
any point of time of crawl.

This makes the crawl on one hand slower, on the other it doesn't block the
database for longer periods of time.

Since the connector can't use any sort of batching anymore this has terrible
performance impact on databases, which are not hosted locally - basically using
Oracle/MS SQL server hosted on a different machine is not an option anymore
because it will become unusable.

There is still one operation which can take longer than three seconds -
downloading all entities from the database in case of full recrawl for deletion
detection (~2 seconds for each 50k users). No idea how to speed this up though,
so I have moved the full recrawl to happen only on Sundays - this needs to
become configurable for countries where Sunday is a working day.

b/7294703