SP Fix Java 7 compiler warnings.
Axis (or our use of it at least) does not support generics in its API.
The contents of lists returned from a MessageElement was not clear to
me, so I have used wildcards everywhere, and left the existing casts
in place. (We would have a cast no matter what, given the lack of
generics in the API).

The SPSiteDataWS.getSiteData method was returning a last modified date
using the current time. I don't think this was intentional. I changed

         sWebMetadata.value.getLastModified().getInstance()

which is equivalent to

         Calendar.getInstance()
to
         sWebMetadata.value.getLastModified()

This will cause the crawler to crawl the site page less often in
metadata-and-URL mode, since it checks the last modified date.

I didn't fix a warning about UriUtil in SPDocumentTest when compiling
using Java 7 without a bootclasspath.

Other changes:

* Removed final from changed lines for parameters and local variables.

* Restored the standard main body in SharePointMain (removing final
  and character string constants).

* When for loops with iterators extended past 80 cols, I changed them
  to while loops rather than wrapping the for loop.

* Fixed indentation error in ListsHelper introduced in r1984.

* Removed some apparently pointless println statements in a test.

Code review: https://codereview.appspot.com/6772045/