Remove roles, part one: move getRoleTypesFor method.
Move the getRoleTypesFor method from Util to AclHelper (the only
caller), and move the SPBasePermissions "class" (a collection of
constants) from SPContants to AclHelper (since getRoleTypesFor is the
only user of the constants). The bodies of the moved method and class
have not been changed, except for a call to new HashSet which is now
Sets.newHashSet to match the other uses.

Other changes:

* Suppress deprecation warnings for RoleType.
* Added final to all the fields in AclHelper.

Code review: https://codereview.appspot.com/12800044/