Fix compiler warnings.
Suppress a deprecation warning on an override of the SAXParser method
getParser that returns a Parser, which is a deprecated interface. Using
the fully-qualified name avoids a deprecation warning on the import.

Restore a deprecation workaround in SharepointClientContext that was
broken in r1594 by "cleaning up" the code. I had originally fixed this
in r1477 to select the undeprecated Protocol constructor by changing
the type of the socket factory argument. The breakage in r1594 also
changed the SSL factory used, from EasySSLProtocolSocketFactory to
SSLProtocolSocketFactory. The gap seems to be that the former works
with self-signed certificates. Since that change has been in place
since 3.0, and is presumably not causing problems, and is presumably
more secure, I'm leaving it alone.

Code review: https://codereview.appspot.com/28180043/